requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via django
astroid==2.6.2
    # via pylint
attrs==21.2.0
    # via pytest
black==21.6b0
    # via -r requirements.in
click==8.0.1
    # via black
django-cors-headers==3.7.0
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==2.4.0
    # via
    #   -r requirements.in
    #   djangorestframework-filters
django==3.2.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework-filters==1.0.0.dev0
    # via -r requirements.in
djangorestframework-simplejwt==4.7.2
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-filters
    #   djangorestframework-simplejwt
factory-boy==3.2.0
    # via -r requirements.in
faker==8.10.1
    # via factory-boy
iniconfig==1.1.1
    # via pytest
isort==5.9.2
    # via pylint
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via pylint
mypy-extensions==0.4.3
    # via black
packaging==21.0
    # via pytest
pathspec==0.8.1
    # via black
pep8==1.7.1
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
psycopg2==2.9.1
    # via -r requirements.in
py==1.10.0
    # via pytest
pyjwt==2.1.0
    # via djangorestframework-simplejwt
pylint==2.9.3
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
pytest-django==4.4.0
    # via -r requirements.in
pytest==6.2.4
    # via
    #   -r requirements.in
    #   pytest-django
python-dateutil==2.8.1
    # via faker
pytz==2021.1
    # via django
regex==2021.7.6
    # via black
six==1.16.0
    # via
    #   -r requirements.in
    #   python-dateutil
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   black
    #   pylint
    #   pytest
wrapt==1.12.1
    # via astroid
